added a segmented target class, and improved bash scripts for running/writing data

+++ scripts/run.sh
@@ -1,8 +1,16 @@
-#! /bin/zsh
+#! /bin/bash
+
+if [ -z "$1" ]; then
+    echo "Usage: $0 <detector_setup>"
+    exit 1
+fi
+
+setup=$1
 
 rm tmp_files/*.csv
 
-./sim run.mac
+./sim "${setup}" run.mac
 
-cd tmp_files
-./merge_csv.sh cd ..
+./merge_csv.sh "${setup}"
+
+rm tmp_files/*thread*.csv
